+++ lib/features/developer/ui/jobs/job_details/ui/widgets/coin_icon_and_salary.dart
