+++ lib/features/developer/ui/jobs/job_details/ui/widgets/job_name_location_type_post_date.dart
